+++ flutter/test/mocks.mocks.dart
@@ -132,8 +132,9 @@ class _FakeSentryTransaction_7 extends _i1.SmartFake
         );
 }
 
-class _FakeMethodCodec_8 extends _i1.SmartFake implements _i4.MethodCodec {
-  _FakeMethodCodec_8(
+class _FakeSdkLifecycleRegistry_8 extends _i1.SmartFake
+    implements _i2.SdkLifecycleRegistry {
+  _FakeSdkLifecycleRegistry_8(
     Object parent,
     Invocation parentInvocation,
   ) : super(
@@ -142,9 +143,19 @@ class _FakeMethodCodec_8 extends _i1.SmartFake implements _i4.MethodCodec {
         );
 }
 
-class _FakeBinaryMessenger_9 extends _i1.SmartFake
+class _FakeMethodCodec_9 extends _i1.SmartFake implements _i4.MethodCodec {
+  _FakeMethodCodec_9(
+    Object parent,
+    Invocation parentInvocation,
+  ) : super(
+          parent,
+          parentInvocation,
+        );
+}
+
+class _FakeBinaryMessenger_10 extends _i1.SmartFake
     implements _i4.BinaryMessenger {
-  _FakeBinaryMessenger_9(
+  _FakeBinaryMessenger_10(
     Object parent,
     Invocation parentInvocation,
   ) : super(
@@ -153,9 +164,9 @@ class _FakeBinaryMessenger_9 extends _i1.SmartFake
         );
 }
 
-class _FakeWidgetsBinding_10 extends _i1.SmartFake
+class _FakeWidgetsBinding_11 extends _i1.SmartFake
     implements _i5.WidgetsBinding {
-  _FakeWidgetsBinding_10(
+  _FakeWidgetsBinding_11(
     Object parent,
     Invocation parentInvocation,
   ) : super(
@@ -164,9 +175,9 @@ class _FakeWidgetsBinding_10 extends _i1.SmartFake
         );
 }
 
-class _FakeSingletonFlutterWindow_11 extends _i1.SmartFake
+class _FakeSingletonFlutterWindow_12 extends _i1.SmartFake
     implements _i6.SingletonFlutterWindow {
-  _FakeSingletonFlutterWindow_11(
+  _FakeSingletonFlutterWindow_12(
     Object parent,
     Invocation parentInvocation,
   ) : super(
@@ -175,9 +186,9 @@ class _FakeSingletonFlutterWindow_11 extends _i1.SmartFake
         );
 }
 
-class _FakePlatformDispatcher_12 extends _i1.SmartFake
+class _FakePlatformDispatcher_13 extends _i1.SmartFake
     implements _i6.PlatformDispatcher {
-  _FakePlatformDispatcher_12(
+  _FakePlatformDispatcher_13(
     Object parent,
     Invocation parentInvocation,
   ) : super(
@@ -186,8 +197,8 @@ class _FakePlatformDispatcher_12 extends _i1.SmartFake
         );
 }
 
-class _FakePointerRouter_13 extends _i1.SmartFake implements _i7.PointerRouter {
-  _FakePointerRouter_13(
+class _FakePointerRouter_14 extends _i1.SmartFake implements _i7.PointerRouter {
+  _FakePointerRouter_14(
     Object parent,
     Invocation parentInvocation,
   ) : super(
@@ -196,9 +207,9 @@ class _FakePointerRouter_13 extends _i1.SmartFake implements _i7.PointerRouter {
         );
 }
 
-class _FakeGestureArenaManager_14 extends _i1.SmartFake
+class _FakeGestureArenaManager_15 extends _i1.SmartFake
     implements _i7.GestureArenaManager {
-  _FakeGestureArenaManager_14(
+  _FakeGestureArenaManager_15(
     Object parent,
     Invocation parentInvocation,
   ) : super(
@@ -207,9 +218,9 @@ class _FakeGestureArenaManager_14 extends _i1.SmartFake
         );
 }
 
-class _FakePointerSignalResolver_15 extends _i1.SmartFake
+class _FakePointerSignalResolver_16 extends _i1.SmartFake
     implements _i7.PointerSignalResolver {
-  _FakePointerSignalResolver_15(
+  _FakePointerSignalResolver_16(
     Object parent,
     Invocation parentInvocation,
   ) : super(
@@ -218,8 +229,8 @@ class _FakePointerSignalResolver_15 extends _i1.SmartFake
         );
 }
 
-class _FakeDuration_16 extends _i1.SmartFake implements Duration {
-  _FakeDuration_16(
+class _FakeDuration_17 extends _i1.SmartFake implements Duration {
+  _FakeDuration_17(
     Object parent,
     Invocation parentInvocation,
   ) : super(
@@ -228,8 +239,8 @@ class _FakeDuration_16 extends _i1.SmartFake implements Duration {
         );
 }
 
-class _FakeSamplingClock_17 extends _i1.SmartFake implements _i7.SamplingClock {
-  _FakeSamplingClock_17(
+class _FakeSamplingClock_18 extends _i1.SmartFake implements _i7.SamplingClock {
+  _FakeSamplingClock_18(
     Object parent,
     Invocation parentInvocation,
   ) : super(
@@ -238,9 +249,9 @@ class _FakeSamplingClock_17 extends _i1.SmartFake implements _i7.SamplingClock {
         );
 }
 
-class _FakeValueNotifier_18<T> extends _i1.SmartFake
+class _FakeValueNotifier_19<T> extends _i1.SmartFake
     implements _i8.ValueNotifier<T> {
-  _FakeValueNotifier_18(
+  _FakeValueNotifier_19(
     Object parent,
     Invocation parentInvocation,
   ) : super(
@@ -249,9 +260,9 @@ class _FakeValueNotifier_18<T> extends _i1.SmartFake
         );
 }
 
-class _FakeHardwareKeyboard_19 extends _i1.SmartFake
+class _FakeHardwareKeyboard_20 extends _i1.SmartFake
     implements _i4.HardwareKeyboard {
-  _FakeHardwareKeyboard_19(
+  _FakeHardwareKeyboard_20(
     Object parent,
     Invocation parentInvocation,
   ) : super(
@@ -260,9 +271,9 @@ class _FakeHardwareKeyboard_19 extends _i1.SmartFake
         );
 }
 
-class _FakeKeyEventManager_20 extends _i1.SmartFake
+class _FakeKeyEventManager_21 extends _i1.SmartFake
     implements _i4.KeyEventManager {
-  _FakeKeyEventManager_20(
+  _FakeKeyEventManager_21(
     Object parent,
     Invocation parentInvocation,
   ) : super(
@@ -271,9 +282,9 @@ class _FakeKeyEventManager_20 extends _i1.SmartFake
         );
 }
 
-class _FakeChannelBuffers_21 extends _i1.SmartFake
+class _FakeChannelBuffers_22 extends _i1.SmartFake
     implements _i6.ChannelBuffers {
-  _FakeChannelBuffers_21(
+  _FakeChannelBuffers_22(
     Object parent,
     Invocation parentInvocation,
   ) : super(
@@ -282,9 +293,9 @@ class _FakeChannelBuffers_21 extends _i1.SmartFake
         );
 }
 
-class _FakeRestorationManager_22 extends _i1.SmartFake
+class _FakeRestorationManager_23 extends _i1.SmartFake
     implements _i4.RestorationManager {
-  _FakeRestorationManager_22(
+  _FakeRestorationManager_23(
     Object parent,
     Invocation parentInvocation,
   ) : super(
@@ -293,8 +304,8 @@ class _FakeRestorationManager_22 extends _i1.SmartFake
         );
 }
 
-class _FakeImageCache_23 extends _i1.SmartFake implements _i9.ImageCache {
-  _FakeImageCache_23(
+class _FakeImageCache_24 extends _i1.SmartFake implements _i9.ImageCache {
+  _FakeImageCache_24(
     Object parent,
     Invocation parentInvocation,
   ) : super(
@@ -303,8 +314,8 @@ class _FakeImageCache_23 extends _i1.SmartFake implements _i9.ImageCache {
         );
 }
 
-class _FakeListenable_24 extends _i1.SmartFake implements _i8.Listenable {
-  _FakeListenable_24(
+class _FakeListenable_25 extends _i1.SmartFake implements _i8.Listenable {
+  _FakeListenable_25(
     Object parent,
     Invocation parentInvocation,
   ) : super(
@@ -313,9 +324,9 @@ class _FakeListenable_24 extends _i1.SmartFake implements _i8.Listenable {
         );
 }
 
-class _FakeAccessibilityFeatures_25 extends _i1.SmartFake
+class _FakeAccessibilityFeatures_26 extends _i1.SmartFake
     implements _i6.AccessibilityFeatures {
-  _FakeAccessibilityFeatures_25(
+  _FakeAccessibilityFeatures_26(
     Object parent,
     Invocation parentInvocation,
   ) : super(
@@ -324,8 +335,8 @@ class _FakeAccessibilityFeatures_25 extends _i1.SmartFake
         );
 }
 
-class _FakeRenderView_26 extends _i1.SmartFake implements _i10.RenderView {
-  _FakeRenderView_26(
+class _FakeRenderView_27 extends _i1.SmartFake implements _i10.RenderView {
+  _FakeRenderView_27(
     Object parent,
     Invocation parentInvocation,
   ) : super(
@@ -338,8 +349,8 @@ class _FakeRenderView_26 extends _i1.SmartFake implements _i10.RenderView {
       super.toString();
 }
 
-class _FakeMouseTracker_27 extends _i1.SmartFake implements _i10.MouseTracker {
-  _FakeMouseTracker_27(
+class _FakeMouseTracker_28 extends _i1.SmartFake implements _i10.MouseTracker {
+  _FakeMouseTracker_28(
     Object parent,
     Invocation parentInvocation,
   ) : super(
@@ -348,9 +359,9 @@ class _FakeMouseTracker_27 extends _i1.SmartFake implements _i10.MouseTracker {
         );
 }
 
-class _FakePlatformMenuDelegate_28 extends _i1.SmartFake
+class _FakePlatformMenuDelegate_29 extends _i1.SmartFake
     implements _i9.PlatformMenuDelegate {
-  _FakePlatformMenuDelegate_28(
+  _FakePlatformMenuDelegate_29(
     Object parent,
     Invocation parentInvocation,
   ) : super(
@@ -359,8 +370,8 @@ class _FakePlatformMenuDelegate_28 extends _i1.SmartFake
         );
 }
 
-class _FakeFocusManager_29 extends _i1.SmartFake implements _i9.FocusManager {
-  _FakeFocusManager_29(
+class _FakeFocusManager_30 extends _i1.SmartFake implements _i9.FocusManager {
+  _FakeFocusManager_30(
     Object parent,
     Invocation parentInvocation,
   ) : super(
@@ -373,8 +384,8 @@ class _FakeFocusManager_29 extends _i1.SmartFake implements _i9.FocusManager {
       super.toString();
 }
 
-class _FakeFuture_30<T1> extends _i1.SmartFake implements _i11.Future<T1> {
-  _FakeFuture_30(
+class _FakeFuture_31<T1> extends _i1.SmartFake implements _i11.Future<T1> {
+  _FakeFuture_31(
     Object parent,
     Invocation parentInvocation,
   ) : super(
@@ -383,8 +394,8 @@ class _FakeFuture_30<T1> extends _i1.SmartFake implements _i11.Future<T1> {
         );
 }
 
-class _FakeCodec_31 extends _i1.SmartFake implements _i6.Codec {
-  _FakeCodec_31(
+class _FakeCodec_32 extends _i1.SmartFake implements _i6.Codec {
+  _FakeCodec_32(
     Object parent,
     Invocation parentInvocation,
   ) : super(
@@ -393,9 +404,9 @@ class _FakeCodec_31 extends _i1.SmartFake implements _i6.Codec {
         );
 }
 
-class _FakeSemanticsHandle_32 extends _i1.SmartFake
+class _FakeSemanticsHandle_33 extends _i1.SmartFake
     implements _i12.SemanticsHandle {
-  _FakeSemanticsHandle_32(
+  _FakeSemanticsHandle_33(
     Object parent,
     Invocation parentInvocation,
   ) : super(
@@ -404,9 +415,9 @@ class _FakeSemanticsHandle_32 extends _i1.SmartFake
         );
 }
 
-class _FakeSemanticsUpdateBuilder_33 extends _i1.SmartFake
+class _FakeSemanticsUpdateBuilder_34 extends _i1.SmartFake
     implements _i6.SemanticsUpdateBuilder {
-  _FakeSemanticsUpdateBuilder_33(
+  _FakeSemanticsUpdateBuilder_34(
     Object parent,
     Invocation parentInvocation,
   ) : super(
@@ -415,9 +426,9 @@ class _FakeSemanticsUpdateBuilder_33 extends _i1.SmartFake
         );
 }
 
-class _FakeViewConfiguration_34 extends _i1.SmartFake
+class _FakeViewConfiguration_35 extends _i1.SmartFake
     implements _i10.ViewConfiguration {
-  _FakeViewConfiguration_34(
+  _FakeViewConfiguration_35(
     Object parent,
     Invocation parentInvocation,
   ) : super(
@@ -426,8 +437,8 @@ class _FakeViewConfiguration_34 extends _i1.SmartFake
         );
 }
 
-class _FakeSceneBuilder_35 extends _i1.SmartFake implements _i6.SceneBuilder {
-  _FakeSceneBuilder_35(
+class _FakeSceneBuilder_36 extends _i1.SmartFake implements _i6.SceneBuilder {
+  _FakeSceneBuilder_36(
     Object parent,
     Invocation parentInvocation,
   ) : super(
@@ -436,9 +447,9 @@ class _FakeSceneBuilder_35 extends _i1.SmartFake implements _i6.SceneBuilder {
         );
 }
 
-class _FakePictureRecorder_36 extends _i1.SmartFake
+class _FakePictureRecorder_37 extends _i1.SmartFake
     implements _i6.PictureRecorder {
-  _FakePictureRecorder_36(
+  _FakePictureRecorder_37(
     Object parent,
     Invocation parentInvocation,
   ) : super(
@@ -447,8 +458,8 @@ class _FakePictureRecorder_36 extends _i1.SmartFake
         );
 }
 
-class _FakeCanvas_37 extends _i1.SmartFake implements _i6.Canvas {
-  _FakeCanvas_37(
+class _FakeCanvas_38 extends _i1.SmartFake implements _i6.Canvas {
+  _FakeCanvas_38(
     Object parent,
     Invocation parentInvocation,
   ) : super(
@@ -457,8 +468,8 @@ class _FakeCanvas_37 extends _i1.SmartFake implements _i6.Canvas {
         );
 }
 
-class _FakeWidget_38 extends _i1.SmartFake implements _i9.Widget {
-  _FakeWidget_38(
+class _FakeWidget_39 extends _i1.SmartFake implements _i9.Widget {
+  _FakeWidget_39(
     Object parent,
     Invocation parentInvocation,
   ) : super(
@@ -471,9 +482,9 @@ class _FakeWidget_38 extends _i1.SmartFake implements _i9.Widget {
       super.toString();
 }
 
-class _FakeSentryFlutterOptions_39 extends _i1.SmartFake
+class _FakeSentryFlutterOptions_40 extends _i1.SmartFake
     implements _i2.SentryFlutterOptions {
-  _FakeSentryFlutterOptions_39(
+  _FakeSentryFlutterOptions_40(
     Object parent,
     Invocation parentInvocation,
   ) : super(
@@ -482,8 +493,8 @@ class _FakeSentryFlutterOptions_39 extends _i1.SmartFake
         );
 }
 
-class _FakeSentryOptions_40 extends _i1.SmartFake implements _i2.SentryOptions {
-  _FakeSentryOptions_40(
+class _FakeSentryOptions_41 extends _i1.SmartFake implements _i2.SentryOptions {
+  _FakeSentryOptions_41(
     Object parent,
     Invocation parentInvocation,
   ) : super(
@@ -492,8 +503,8 @@ class _FakeSentryOptions_40 extends _i1.SmartFake implements _i2.SentryOptions {
         );
 }
 
-class _FakeScope_41 extends _i1.SmartFake implements _i2.Scope {
-  _FakeScope_41(
+class _FakeScope_42 extends _i1.SmartFake implements _i2.Scope {
+  _FakeScope_42(
     Object parent,
     Invocation parentInvocation,
   ) : super(
@@ -502,8 +513,8 @@ class _FakeScope_41 extends _i1.SmartFake implements _i2.Scope {
         );
 }
 
-class _FakeHub_42 extends _i1.SmartFake implements _i2.Hub {
-  _FakeHub_42(
+class _FakeHub_43 extends _i1.SmartFake implements _i2.Hub {
+  _FakeHub_43(
     Object parent,
     Invocation parentInvocation,
   ) : super(
@@ -1575,6 +1586,15 @@ class MockSentryClient extends _i1.Mock implements _i2.SentryClient {
   }
 
   @override
+  _i2.SdkLifecycleRegistry get lifeCycleRegistry => (super.noSuchMethod(
+        Invocation.getter(#lifeCycleRegistry),
+        returnValue: _FakeSdkLifecycleRegistry_8(
+          this,
+          Invocation.getter(#lifeCycleRegistry),
+        ),
+      ) as _i2.SdkLifecycleRegistry);
+
+  @override
   _i11.Future<_i2.SentryId> captureEvent(
     _i2.SentryEvent? event, {
     _i2.Scope? scope,
@@ -1761,17 +1781,6 @@ class MockSentryClient extends _i1.Mock implements _i2.SentryClient {
         ),
         returnValueForMissingStub: null,
       );
-
-  @override
-  void registerCallback<T extends _i2.SdkLifecycleEvent>(
-          _i2.SdkLifecycleCallback<T>? callback) =>
-      super.noSuchMethod(
-        Invocation.method(
-          #registerCallback,
-          [callback],
-        ),
-        returnValueForMissingStub: null,
-      );
 }
 
 /// A class which mocks [MethodChannel].
@@ -1794,7 +1803,7 @@ class MockMethodChannel extends _i1.Mock implements _i4.MethodChannel {
   @override
   _i4.MethodCodec get codec => (super.noSuchMethod(
         Invocation.getter(#codec),
-        returnValue: _FakeMethodCodec_8(
+        returnValue: _FakeMethodCodec_9(
           this,
           Invocation.getter(#codec),
         ),
@@ -1803,7 +1812,7 @@ class MockMethodChannel extends _i1.Mock implements _i4.MethodChannel {
   @override
   _i4.BinaryMessenger get binaryMessenger => (super.noSuchMethod(
         Invocation.getter(#binaryMessenger),
-        returnValue: _FakeBinaryMessenger_9(
+        returnValue: _FakeBinaryMessenger_10(
           this,
           Invocation.getter(#binaryMessenger),
         ),
@@ -2154,7 +2163,7 @@ class MockBindingWrapper extends _i1.Mock implements _i2.BindingWrapper {
           #ensureInitialized,
           [],
         ),
-        returnValue: _FakeWidgetsBinding_10(
+        returnValue: _FakeWidgetsBinding_11(
           this,
           Invocation.method(
             #ensureInitialized,
@@ -2176,7 +2185,7 @@ class MockWidgetsFlutterBinding extends _i1.Mock
   @override
   _i6.SingletonFlutterWindow get window => (super.noSuchMethod(
         Invocation.getter(#window),
-        returnValue: _FakeSingletonFlutterWindow_11(
+        returnValue: _FakeSingletonFlutterWindow_12(
           this,
           Invocation.getter(#window),
         ),
@@ -2185,7 +2194,7 @@ class MockWidgetsFlutterBinding extends _i1.Mock
   @override
   _i6.PlatformDispatcher get platformDispatcher => (super.noSuchMethod(
         Invocation.getter(#platformDispatcher),
-        returnValue: _FakePlatformDispatcher_12(
+        returnValue: _FakePlatformDispatcher_13(
           this,
           Invocation.getter(#platformDispatcher),
         ),
@@ -2200,7 +2209,7 @@ class MockWidgetsFlutterBinding extends _i1.Mock
   @override
   _i7.PointerRouter get pointerRouter => (super.noSuchMethod(
         Invocation.getter(#pointerRouter),
-        returnValue: _FakePointerRouter_13(
+        returnValue: _FakePointerRouter_14(
           this,
           Invocation.getter(#pointerRouter),
         ),
@@ -2209,7 +2218,7 @@ class MockWidgetsFlutterBinding extends _i1.Mock
   @override
   _i7.GestureArenaManager get gestureArena => (super.noSuchMethod(
         Invocation.getter(#gestureArena),
-        returnValue: _FakeGestureArenaManager_14(
+        returnValue: _FakeGestureArenaManager_15(
           this,
           Invocation.getter(#gestureArena),
         ),
@@ -2218,7 +2227,7 @@ class MockWidgetsFlutterBinding extends _i1.Mock
   @override
   _i7.PointerSignalResolver get pointerSignalResolver => (super.noSuchMethod(
         Invocation.getter(#pointerSignalResolver),
-        returnValue: _FakePointerSignalResolver_15(
+        returnValue: _FakePointerSignalResolver_16(
           this,
           Invocation.getter(#pointerSignalResolver),
         ),
@@ -2233,7 +2242,7 @@ class MockWidgetsFlutterBinding extends _i1.Mock
   @override
   Duration get samplingOffset => (super.noSuchMethod(
         Invocation.getter(#samplingOffset),
-        returnValue: _FakeDuration_16(
+        returnValue: _FakeDuration_17(
           this,
           Invocation.getter(#samplingOffset),
         ),
@@ -2242,7 +2251,7 @@ class MockWidgetsFlutterBinding extends _i1.Mock
   @override
   _i7.SamplingClock get samplingClock => (super.noSuchMethod(
         Invocation.getter(#samplingClock),
-        returnValue: _FakeSamplingClock_17(
+        returnValue: _FakeSamplingClock_18(
           this,
           Invocation.getter(#samplingClock),
         ),
@@ -2309,7 +2318,7 @@ class MockWidgetsFlutterBinding extends _i1.Mock
   @override
   Duration get currentFrameTimeStamp => (super.noSuchMethod(
         Invocation.getter(#currentFrameTimeStamp),
-        returnValue: _FakeDuration_16(
+        returnValue: _FakeDuration_17(
           this,
           Invocation.getter(#currentFrameTimeStamp),
         ),
@@ -2318,7 +2327,7 @@ class MockWidgetsFlutterBinding extends _i1.Mock
   @override
   Duration get currentSystemFrameTimeStamp => (super.noSuchMethod(
         Invocation.getter(#currentSystemFrameTimeStamp),
-        returnValue: _FakeDuration_16(
+        returnValue: _FakeDuration_17(
           this,
           Invocation.getter(#currentSystemFrameTimeStamp),
         ),
@@ -2337,7 +2346,7 @@ class MockWidgetsFlutterBinding extends _i1.Mock
   @override
   _i8.ValueNotifier<int?> get accessibilityFocus => (super.noSuchMethod(
         Invocation.getter(#accessibilityFocus),
-        returnValue: _FakeValueNotifier_18<int?>(
+        returnValue: _FakeValueNotifier_19<int?>(
           this,
           Invocation.getter(#accessibilityFocus),
         ),
@@ -2346,7 +2355,7 @@ class MockWidgetsFlutterBinding extends _i1.Mock
   @override
   _i4.HardwareKeyboard get keyboard => (super.noSuchMethod(
         Invocation.getter(#keyboard),
-        returnValue: _FakeHardwareKeyboard_19(
+        returnValue: _FakeHardwareKeyboard_20(
           this,
           Invocation.getter(#keyboard),
         ),
@@ -2355,7 +2364,7 @@ class MockWidgetsFlutterBinding extends _i1.Mock
   @override
   _i4.KeyEventManager get keyEventManager => (super.noSuchMethod(
         Invocation.getter(#keyEventManager),
-        returnValue: _FakeKeyEventManager_20(
+        returnValue: _FakeKeyEventManager_21(
           this,
           Invocation.getter(#keyEventManager),
         ),
@@ -2364,7 +2373,7 @@ class MockWidgetsFlutterBinding extends _i1.Mock
   @override
   _i4.BinaryMessenger get defaultBinaryMessenger => (super.noSuchMethod(
         Invocation.getter(#defaultBinaryMessenger),
-        returnValue: _FakeBinaryMessenger_9(
+        returnValue: _FakeBinaryMessenger_10(
           this,
           Invocation.getter(#defaultBinaryMessenger),
         ),
@@ -2373,7 +2382,7 @@ class MockWidgetsFlutterBinding extends _i1.Mock
   @override
   _i6.ChannelBuffers get channelBuffers => (super.noSuchMethod(
         Invocation.getter(#channelBuffers),
-        returnValue: _FakeChannelBuffers_21(
+        returnValue: _FakeChannelBuffers_22(
           this,
           Invocation.getter(#channelBuffers),
         ),
@@ -2382,7 +2391,7 @@ class MockWidgetsFlutterBinding extends _i1.Mock
   @override
   _i4.RestorationManager get restorationManager => (super.noSuchMethod(
         Invocation.getter(#restorationManager),
-        returnValue: _FakeRestorationManager_22(
+        returnValue: _FakeRestorationManager_23(
           this,
           Invocation.getter(#restorationManager),
         ),
@@ -2391,7 +2400,7 @@ class MockWidgetsFlutterBinding extends _i1.Mock
   @override
   _i9.ImageCache get imageCache => (super.noSuchMethod(
         Invocation.getter(#imageCache),
-        returnValue: _FakeImageCache_23(
+        returnValue: _FakeImageCache_24(
           this,
           Invocation.getter(#imageCache),
         ),
@@ -2400,7 +2409,7 @@ class MockWidgetsFlutterBinding extends _i1.Mock
   @override
   _i8.Listenable get systemFonts => (super.noSuchMethod(
         Invocation.getter(#systemFonts),
-        returnValue: _FakeListenable_24(
+        returnValue: _FakeListenable_25(
           this,
           Invocation.getter(#systemFonts),
         ),
@@ -2421,7 +2430,7 @@ class MockWidgetsFlutterBinding extends _i1.Mock
   @override
   _i6.AccessibilityFeatures get accessibilityFeatures => (super.noSuchMethod(
         Invocation.getter(#accessibilityFeatures),
-        returnValue: _FakeAccessibilityFeatures_25(
+        returnValue: _FakeAccessibilityFeatures_26(
           this,
           Invocation.getter(#accessibilityFeatures),
         ),
@@ -2445,7 +2454,7 @@ class MockWidgetsFlutterBinding extends _i1.Mock
   @override
   _i10.RenderView get renderView => (super.noSuchMethod(
         Invocation.getter(#renderView),
-        returnValue: _FakeRenderView_26(
+        returnValue: _FakeRenderView_27(
           this,
           Invocation.getter(#renderView),
         ),
@@ -2454,7 +2463,7 @@ class MockWidgetsFlutterBinding extends _i1.Mock
   @override
   _i10.MouseTracker get mouseTracker => (super.noSuchMethod(
         Invocation.getter(#mouseTracker),
-        returnValue: _FakeMouseTracker_27(
+        returnValue: _FakeMouseTracker_28(
           this,
           Invocation.getter(#mouseTracker),
         ),
@@ -2484,7 +2493,7 @@ class MockWidgetsFlutterBinding extends _i1.Mock
   @override
   _i9.PlatformMenuDelegate get platformMenuDelegate => (super.noSuchMethod(
         Invocation.getter(#platformMenuDelegate),
-        returnValue: _FakePlatformMenuDelegate_28(
+        returnValue: _FakePlatformMenuDelegate_29(
           this,
           Invocation.getter(#platformMenuDelegate),
         ),
@@ -2506,7 +2515,7 @@ class MockWidgetsFlutterBinding extends _i1.Mock
   _i8.ValueNotifier<bool> get debugShowWidgetInspectorOverrideNotifier =>
       (super.noSuchMethod(
         Invocation.getter(#debugShowWidgetInspectorOverrideNotifier),
-        returnValue: _FakeValueNotifier_18<bool>(
+        returnValue: _FakeValueNotifier_19<bool>(
           this,
           Invocation.getter(#debugShowWidgetInspectorOverrideNotifier),
         ),
@@ -2516,7 +2525,7 @@ class MockWidgetsFlutterBinding extends _i1.Mock
   _i8.ValueNotifier<bool> get debugWidgetInspectorSelectionOnTapEnabled =>
       (super.noSuchMethod(
         Invocation.getter(#debugWidgetInspectorSelectionOnTapEnabled),
-        returnValue: _FakeValueNotifier_18<bool>(
+        returnValue: _FakeValueNotifier_19<bool>(
           this,
           Invocation.getter(#debugWidgetInspectorSelectionOnTapEnabled),
         ),
@@ -2525,7 +2534,7 @@ class MockWidgetsFlutterBinding extends _i1.Mock
   @override
   _i9.FocusManager get focusManager => (super.noSuchMethod(
         Invocation.getter(#focusManager),
-        returnValue: _FakeFocusManager_29(
+        returnValue: _FakeFocusManager_30(
           this,
           Invocation.getter(#focusManager),
         ),
@@ -2925,7 +2934,7 @@ class MockWidgetsFlutterBinding extends _i1.Mock
               ),
               (T v) => _i11.Future<T>.value(v),
             ) ??
-            _FakeFuture_30<T>(
+            _FakeFuture_31<T>(
               this,
               Invocation.method(
                 #scheduleTask,
@@ -3115,7 +3124,7 @@ class MockWidgetsFlutterBinding extends _i1.Mock
           #createBinaryMessenger,
           [],
         ),
-        returnValue: _FakeBinaryMessenger_9(
+        returnValue: _FakeBinaryMessenger_10(
           this,
           Invocation.method(
             #createBinaryMessenger,
@@ -3214,7 +3223,7 @@ class MockWidgetsFlutterBinding extends _i1.Mock
           #createRestorationManager,
           [],
         ),
-        returnValue: _FakeRestorationManager_22(
+        returnValue: _FakeRestorationManager_23(
           this,
           Invocation.method(
             #createRestorationManager,
@@ -3249,7 +3258,7 @@ class MockWidgetsFlutterBinding extends _i1.Mock
           #createImageCache,
           [],
         ),
-        returnValue: _FakeImageCache_23(
+        returnValue: _FakeImageCache_24(
           this,
           Invocation.method(
             #createImageCache,
@@ -3275,7 +3284,7 @@ class MockWidgetsFlutterBinding extends _i1.Mock
             #allowUpscaling: allowUpscaling,
           },
         ),
-        returnValue: _i11.Future<_i6.Codec>.value(_FakeCodec_31(
+        returnValue: _i11.Future<_i6.Codec>.value(_FakeCodec_32(
           this,
           Invocation.method(
             #instantiateImageCodecFromBuffer,
@@ -3300,7 +3309,7 @@ class MockWidgetsFlutterBinding extends _i1.Mock
           [buffer],
           {#getTargetSize: getTargetSize},
         ),
-        returnValue: _i11.Future<_i6.Codec>.value(_FakeCodec_31(
+        returnValue: _i11.Future<_i6.Codec>.value(_FakeCodec_32(
           this,
           Invocation.method(
             #instantiateImageCodecWithSize,
@@ -3358,7 +3367,7 @@ class MockWidgetsFlutterBinding extends _i1.Mock
           #ensureSemantics,
           [],
         ),
-        returnValue: _FakeSemanticsHandle_32(
+        returnValue: _FakeSemanticsHandle_33(
           this,
           Invocation.method(
             #ensureSemantics,
@@ -3393,7 +3402,7 @@ class MockWidgetsFlutterBinding extends _i1.Mock
           #createSemanticsUpdateBuilder,
           [],
         ),
-        returnValue: _FakeSemanticsUpdateBuilder_33(
+        returnValue: _FakeSemanticsUpdateBuilder_34(
           this,
           Invocation.method(
             #createSemanticsUpdateBuilder,
@@ -3443,7 +3452,7 @@ class MockWidgetsFlutterBinding extends _i1.Mock
           #createViewConfigurationFor,
           [renderView],
         ),
-        returnValue: _FakeViewConfiguration_34(
+        returnValue: _FakeViewConfiguration_35(
           this,
           Invocation.method(
             #createViewConfigurationFor,
@@ -3458,7 +3467,7 @@ class MockWidgetsFlutterBinding extends _i1.Mock
           #createSceneBuilder,
           [],
         ),
-        returnValue: _FakeSceneBuilder_35(
+        returnValue: _FakeSceneBuilder_36(
           this,
           Invocation.method(
             #createSceneBuilder,
@@ -3473,7 +3482,7 @@ class MockWidgetsFlutterBinding extends _i1.Mock
           #createPictureRecorder,
           [],
         ),
-        returnValue: _FakePictureRecorder_36(
+        returnValue: _FakePictureRecorder_37(
           this,
           Invocation.method(
             #createPictureRecorder,
@@ -3488,7 +3497,7 @@ class MockWidgetsFlutterBinding extends _i1.Mock
           #createCanvas,
           [recorder],
         ),
-        returnValue: _FakeCanvas_37(
+        returnValue: _FakeCanvas_38(
           this,
           Invocation.method(
             #createCanvas,
@@ -3639,7 +3648,7 @@ class MockWidgetsFlutterBinding extends _i1.Mock
           #wrapWithDefaultView,
           [rootWidget],
         ),
-        returnValue: _FakeWidget_38(
+        returnValue: _FakeWidget_39(
           this,
           Invocation.method(
             #wrapWithDefaultView,
@@ -3767,7 +3776,7 @@ class MockTimeToDisplayTracker extends _i1.Mock
   @override
   _i2.SentryFlutterOptions get options => (super.noSuchMethod(
         Invocation.getter(#options),
-        returnValue: _FakeSentryFlutterOptions_39(
+        returnValue: _FakeSentryFlutterOptions_40(
           this,
           Invocation.getter(#options),
         ),
@@ -3945,7 +3954,7 @@ class MockHub extends _i1.Mock implements _i2.Hub {
   @override
   _i2.SentryOptions get options => (super.noSuchMethod(
         Invocation.getter(#options),
-        returnValue: _FakeSentryOptions_40(
+        returnValue: _FakeSentryOptions_41(
           this,
           Invocation.getter(#options),
         ),
@@ -3969,11 +3978,17 @@ class MockHub extends _i1.Mock implements _i2.Hub {
   @override
   _i2.Scope get scope => (super.noSuchMethod(
         Invocation.getter(#scope),
-        returnValue: _FakeScope_41(
+        returnValue: _FakeScope_42(
           this,
           Invocation.getter(#scope),
         ),
       ) as _i2.Scope);
+
+  @override
+  Map<Type, List<Function>> get lifecycleCallbacks => (super.noSuchMethod(
+        Invocation.getter(#lifecycleCallbacks),
+        returnValue: <Type, List<Function>>{},
+      ) as Map<Type, List<Function>>);
 
   @override
   set profilerFactory(_i15.SentryProfilerFactory? value) => super.noSuchMethod(
@@ -4151,7 +4166,7 @@ class MockHub extends _i1.Mock implements _i2.Hub {
           #clone,
           [],
         ),
-        returnValue: _FakeHub_42(
+        returnValue: _FakeHub_43(
           this,
           Invocation.method(
             #clone,
@@ -4321,11 +4336,22 @@ class MockHub extends _i1.Mock implements _i2.Hub {
       );
 
   @override
-  void registerCallback<T extends _i2.SdkLifecycleEvent>(
+  void registerSdkLifecycleCallback<T extends _i2.SdkLifecycleEvent>(
           _i2.SdkLifecycleCallback<T>? callback) =>
       super.noSuchMethod(
         Invocation.method(
-          #registerCallback,
+          #registerSdkLifecycleCallback,
+          [callback],
+        ),
+        returnValueForMissingStub: null,
+      );
+
+  @override
+  void removeSdkLifecycleCallback<T extends _i2.SdkLifecycleEvent>(
+          _i2.SdkLifecycleCallback<T>? callback) =>
+      super.noSuchMethod(
+        Invocation.method(
+          #removeSdkLifecycleCallback,
           [callback],
         ),
         returnValueForMissingStub: null,